Split project up into utility and settings files
