+++ settings.py
@@ -1,0 +1,8 @@
+# Lets try to keep a revision tracking via commit number.
+revision = 'c+89'
+
+# Set this for parameters checking
+hypedSkus = ['BY9612', 'BY1605', 'BY9611']
+
+# Code to indicate a shitty exit from the script
+exitCode = 1
